requirements.txt
alabaster==0.7.12
asgiref==3.3.4
Babel==2.9.0
certifi==2020.12.5
distlib==0.3.1
Django==3.2
django-filter==2.4.0
djangorestframework==3.12.4
docutils==0.17
imagesize==1.2.0
jedi==0.18.0
Markdown==3.3.4
numpydoc==1.1.0
packaging==20.9
parso==0.8.2
psycopg2==2.8.4
Pygments==2.8.1
pytz==2021.1
snowballstemmer==2.1.0
Sphinx==3.5.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
